Spring: Make id's long again

+++ spring-portfolio/src/main/java/com/portfolio/HostnameRepository.java
@@ -7,7 +7,7 @@ import com.portfolio.Hostname;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface HostnameRepository extends CrudRepository<Hostname, Integer> {
+public interface HostnameRepository extends CrudRepository<Hostname, Long> {
 
     public Hostname findByName(String name);
 }
